refactor(web): move ping component to separate route and clean up unused code
